+++ server/src/AnalysisManager.ts
@@ -5,35 +5,66 @@ import * as Parser from "web-tree-sitter";
 import {TamarinSymbolTable,createSymbolTable} from "./symbol_table/create_symbol_table";
 import { checks_with_table } from "./features/checks/index";
 import { DeclarationType} from "./symbol_table/tamarinTypes";
-//TODO : REFACTORING
+import * as fs from 'fs';
+import * as path from 'path';
+import { URI } from 'vscode-uri'; 
+
 export class AnalysisManager{
-    private parser: Parser|undefined;
-    private symbolTable: Map<string, TamarinSymbolTable>;
+    private spthyParser: Parser|undefined;
+    private splibParser: Parser|undefined;
+    private documentSymbolTables: Map<string, TamarinSymbolTable>;
 
     constructor() {
-        this.symbolTable = new Map<string, TamarinSymbolTable>();
+        this.documentSymbolTables = new Map<string, TamarinSymbolTable>();
     }
 
-    public async initParser(parserPath: string): Promise<void> {
+    public async initParsers(SpthyParserPath: string, SplibParserPath: string): Promise<void> {
         await Parser.init();
-        this.parser = new Parser();
-        const Tamarin = await Parser.Language.load(parserPath);
-        this.parser.setLanguage(Tamarin);
-        console.log("Parser initialized with Tamarin language.");
+
+        this.spthyParser = new Parser();
+        const spthyLang = await Parser.Language.load(SpthyParserPath);
+        this.spthyParser.setLanguage(spthyLang);
+        console.log("SPTHY Parser initialized with Tamarin language.");
+    
+        this.splibParser = new Parser();
+        const splibLang = await Parser.Language.load(SplibParserPath);
+        this.splibParser.setLanguage(splibLang);
+        console.log("SPLIB Parser initialized with Tamarin language.");
+
     }
+
+
     //error and warning display
     public async AnalyseDocument(document: TextDocument): Promise<Diagnostic[]> {
-        if (!this.parser) {
+        if (!this.spthyParser || !this.splibParser) {
             throw new Error("Parser not initialized");
         }
-        const tree =  this.parser.parse(document.getText());
+        const newSymbolTables = new Map<string, TamarinSymbolTable>();
+        const visitedUrisInThisPass = new Set<string>();
+        await this.AnalyseImports(document.uri, document.getText(), newSymbolTables, visitedUrisInThisPass);
+        this.documentSymbolTables = newSymbolTables;
+        console.log(`[AnalysisManager] Cache updated. Total files analyzed: ${this.documentSymbolTables.size}`);
+
+        const parserToUse: Parser = document.uri.endsWith('.splib')
+            ? this.splibParser
+            : this.spthyParser;
+
+        if (!parserToUse) {
+            throw new Error(`No parser available for URI: ${document.uri}`);
+        }
+
+        const tree = parserToUse.parse(document.getText());
+
         const {diagnostics: syntaxDiagnostics } = await detect_errors(tree.rootNode,document);
         const { symbolTable, diags: symbolTableChecks } = await createSymbolTable(tree.rootNode, document);
-        this.symbolTable.set(document.uri, symbolTable);
-        const wellformednessDiagnostics = await checks_with_table(symbolTable, document, tree.rootNode);
+        this.documentSymbolTables.set(document.uri, symbolTable);
+        // checks with table
+
+        const wellformednessDiagnostics = await checks_with_table(symbolTable, document, tree.rootNode,this.documentSymbolTables);
         console.log("Symbol table created for:", document.uri);
         console.log("Number of symbols found:", symbolTable.getSymbols().length);
         console.log("Wellformedness diagnostics:", wellformednessDiagnostics.length);
+        console.log("Symbol table", symbolTable);
         const allDiagnostics = [
             ...syntaxDiagnostics,
             ...wellformednessDiagnostics,
@@ -41,17 +72,55 @@ export class AnalysisManager{
         ];
         return allDiagnostics;
     }
+
+    public clearCache(): void {
+        this.documentSymbolTables.clear();
+    }
+    
+    private async AnalyseImports(uri: string, 
+        content: string | undefined, 
+        tablesForThisPass: Map<string, TamarinSymbolTable>, 
+        visitedInThisPass: Set<string>,
+        ): Promise<void> {
+        if (!this.spthyParser || !this.splibParser) return;
+        if (visitedInThisPass.has(uri)) {
+            return;
+        }
+        visitedInThisPass.add(uri);
+        let fileContent = content;
+        if (fileContent === undefined) {
+            try {
+                const fsPath = URI.parse(uri).fsPath;
+                fileContent = fs.readFileSync(fsPath, 'utf8');
+            } catch (error) {
+                console.error(`[AnalysisManager] Failed to read imported file ${uri}: ${error}`);
+                return;
+            }
+        }
+
+        const doc = TextDocument.create(uri, 'tamarin', 1, fileContent);
+        const tree = this.splibParser.parse(doc.getText());
+        const { symbolTable} = await createSymbolTable(tree.rootNode, doc);
+        tablesForThisPass.set(uri, symbolTable);
+        const includePaths = symbolTable.getIncludes() || [];
+        const currentDir = path.dirname(URI.parse(uri).fsPath);
+        for (const relativePath of includePaths) {
+            const absolutePath = path.resolve(currentDir, relativePath);
+            const importedUri = URI.file(absolutePath).toString();
+            await this.AnalyseImports(importedUri, undefined, tablesForThisPass, visitedInThisPass);
+        }
+    }
     //goto
     public getDefinition(document: TextDocument, position: Position): Location | null {
-        if(!this.parser) {
+        if(!this.spthyParser) {
             throw new Error("Parser not initialized");
         }
-        const table = this.symbolTable.get(document.uri);
+        const table = this.documentSymbolTables.get(document.uri);
         if (!table) {
             console.error(`No symbol table found for document: ${document.uri}`);
             return null;
         }
-        const tree= this.parser.parse(document.getText());
+        const tree= this.spthyParser.parse(document.getText());
         const point = {row: position.line, column: position.character};
         const nodeAtcursor = tree.rootNode.descendantForPosition(point);
         if (!nodeAtcursor) {
@@ -74,16 +143,16 @@ export class AnalysisManager{
 
     public handleDocumentClose(uri: string): void {
         console.log(`Document closed: ${uri}. Cleaning up symbol table.`);
-        this.symbolTable.delete(uri);
+        this.documentSymbolTables.delete(uri);
     }
     //rename
     public handleRenameRequest(document: TextDocument, position: Position, newName: string): WorkspaceEdit | null {
-        const table = this.symbolTable.get(document.uri);
+        const table = this.documentSymbolTables.get(document.uri);
         if (!table) return null
-        if (!this.parser) {
+        if (!this.spthyParser) {
             throw new Error("Parser not initialized");
         }
-        const tree = this.parser.parse(document.getText());
+        const tree = this.spthyParser.parse(document.getText());
         const point = {row: position.line, column: position.character};
         const nodeAtCursor = tree.rootNode.descendantForPosition(point);
         if (!nodeAtCursor) {