+++ server/src/features/checks/checkVariableTypes.ts
@@ -4,15 +4,22 @@ import { TextDocument } from "vscode-languageserver-textdocument";
 import { Diagnostic} from 'vscode-languageserver';
 import { DeclarationType} from "../../symbol_table/tamarinTypes";
 
+const variableDeclarationTypes = [
+    DeclarationType.PRVariable,
+    DeclarationType.CCLVariable,
+    DeclarationType.ActionFVariable,
+    DeclarationType.LemmaVariable,
+];
+
 /* Function used to check if all variables with the same name  in a rule have the same type,
 Also checks that if a variables is in the right side of a macro or equation it is also present in the left side  */
 export function check_variables_type_is_consistent_inside_a_rule(symbol_table : TamarinSymbolTable, editor: TextDocument) : Diagnostic[]{
     const diags : Diagnostic[] = [];
     for (let i = 0 ; i < symbol_table.getSymbols().length; i++){
         const current_symbol = symbol_table.getSymbol(i);
-        if(current_symbol.declaration === DeclarationType.PRVariable || current_symbol.declaration === DeclarationType.CCLVariable || current_symbol.declaration === DeclarationType.ActionFVariable ){
+        if(variableDeclarationTypes.includes(current_symbol.declaration)){
             for(let j = 0; j < symbol_table.getSymbols().length; j++){
-                if((current_symbol.declaration === DeclarationType.PRVariable || current_symbol.declaration === DeclarationType.CCLVariable || current_symbol.declaration === DeclarationType.ActionFVariable) && i !== j){
+                if(variableDeclarationTypes.includes(symbol_table.getSymbol(j).declaration) && i !== j){
                     if(current_symbol.context === symbol_table.getSymbol(j).context && current_symbol.name === symbol_table.getSymbol(j).name){
                         if(current_symbol.type === symbol_table.getSymbol(j).type){
                             continue;