+++ server/src/features/tests/analysisManager.test.ts
@@ -9,8 +9,9 @@ const docUri = 'file:///test.spthy';
 
 beforeAll(async () => {
     analysisManager = new AnalysisManager();
-    const parserPath = path.resolve(process.cwd(), 'server/grammar/tree-sitter-tamarin/tree-sitter-spthy.wasm');
-    await analysisManager.initParser(parserPath);
+    const SpthyParserPath = path.resolve(process.cwd(), 'server/grammar/tree-sitter-tamarin/tree-sitter-spthy.wasm');
+    const SplibParserPath = path.resolve(process.cwd(), 'server/grammar/tree-sitter-tamarin/tree-sitter-splib.wasm');
+    await analysisManager.initParsers(SpthyParserPath,SplibParserPath);
 });
 
 const setupTest = async (content: string): Promise<TextDocument> => {