+++ server/src/server.ts
@@ -1,4 +1,3 @@
-// Dans server/src/server.ts
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import {
     createConnection,
@@ -23,9 +22,10 @@ let analysisManager: AnalysisManager;
 
 
 connection.onInitialize(async (params: InitializeParams) => {
-    const parserPath = params.initializationOptions.parserPath;
+    const SpthyParserPath = params.initializationOptions.SpthyParserPath;
+    const SplibParserPath = params.initializationOptions.SplibParserPath;
     analysisManager = new AnalysisManager();
-    await analysisManager.initParser(parserPath)
+    await analysisManager.initParsers(SpthyParserPath,SplibParserPath)
     console.error('[Server] Received "initialize" request from client.');
     const capabilities: ServerCapabilities = {
         textDocumentSync: TextDocumentSyncKind.Full,